[FEAT] improve tickets manager and how tickets are displayed on the index + query status to index ticket API
