+++ resources/js/pages/tenants/assets/show.tsx
@@ -1,20 +1,12 @@
 import { DocumentManager } from '@/components/tenant/documentManager';
-
 import { PictureManager } from '@/components/tenant/pictureManager';
+import { TicketManager } from '@/components/tenant/ticketManager';
 import { Button } from '@/components/ui/button';
-import { Checkbox } from '@/components/ui/checkbox';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Table, TableBody, TableBodyData, TableBodyRow, TableHead, TableHeadData, TableHeadRow } from '@/components/ui/table';
-import { Textarea } from '@/components/ui/textarea';
 import AppLayout from '@/layouts/app-layout';
-import { Asset, Ticket, type BreadcrumbItem } from '@/types';
-import { Head, useForm, usePage } from '@inertiajs/react';
-import axios from 'axios';
-import { FormEventHandler, useState } from 'react';
+import { Asset, type BreadcrumbItem } from '@/types';
+import { Head, useForm } from '@inertiajs/react';
 
 export default function ShowAsset({ asset }: { asset: Asset }) {
-    const auth = usePage().props.auth.user;
     const breadcrumbs: BreadcrumbItem[] = [
         {
             title: `${asset.reference_code} - ${asset.maintainable.name}`,
@@ -28,155 +20,6 @@ export default function ShowAsset({ asset }: { asset: Asset }) {
         destroy(route(`tenant.assets.destroy`, asset.code));
     };
 
-    const closeTicket = async (id: number) => {
-        try {
-            const response = await axios.patch(route('api.tickets.close', id));
-            if (response.data.status === 'success') {
-                fetchTickets();
-            }
-        } catch (error) {
-            console.error('Erreur lors de la suppression', error);
-        }
-    };
-
-    const fetchTickets = async () => {
-        try {
-            const response = await axios.get(`/api/v1/assets/${asset.code}/tickets`);
-            setTickets(await response.data);
-        } catch (error) {
-            console.error('Erreur lors de la recherche :', error);
-        }
-    };
-
-    const [addTicketModal, setAddTicketModal] = useState<boolean>(false);
-    const [submitTypeTicket, setSubmitTypeTicket] = useState<'edit' | 'new'>('edit');
-    const updateTicketData = {
-        ticket_id: 0,
-        location_type: 'assets',
-        location_id: asset.id,
-        being_notified: false,
-        description: '',
-        reported_by: auth.id,
-        reporter_email: auth.email,
-        pictures: [],
-    };
-
-    const submitEditTicket: FormEventHandler = async (e) => {
-        e.preventDefault();
-        try {
-            const response = await axios.patch(route('api.tickets.update', newTicketData.ticket_id), newTicketData);
-            fetchTickets();
-            closeModalTicket();
-            // }
-        } catch (error) {
-            console.log(error);
-        }
-    };
-
-    const submitNewTicket: FormEventHandler = async (e) => {
-        e.preventDefault();
-        try {
-            const response = await axios.post(route('api.tickets.store'), newTicketData, {
-                headers: {
-                    'Content-Type': 'multipart/form-data',
-                },
-            });
-            if (response.data.status === 'success') {
-                fetchTickets();
-                closeModalTicket();
-            }
-        } catch (error) {
-            console.log(error);
-        }
-    };
-
-    const [newTicketData, setNewTicketData] = useState<FormDataTicket>(updateTicketData);
-
-    const closeModalTicket = () => {
-        setAddTicketModal(false);
-        setNewTicketData(updateTicketData);
-        setSubmitTypeTicket('edit');
-    };
-
-    const [tickets, setTickets] = useState<Ticket[]>(asset.tickets);
-
-    const editTicket = async (id: number) => {
-        setSubmitTypeTicket('edit');
-        try {
-            const response = await axios.get(route('api.tickets.get', id), {});
-            setNewTicketData((prev) => ({
-                ...prev,
-                ticket_id: response.data.data.id,
-                description: response.data.data.description,
-                being_notified: response.data?.data.being_notified,
-            }));
-
-            setAddTicketModal(true);
-
-            // }
-        } catch (error) {
-            console.log(error);
-        }
-    };
-
-    const addTicket = () => {
-        return (
-            <div className="bg-background/50 absolute inset-0 z-50">
-                <div className="bg-background/20 flex h-dvh items-center justify-center">
-                    <div className="bg-background flex items-center justify-center p-10">
-                        <form onSubmit={submitTypeTicket === 'edit' ? submitEditTicket : submitNewTicket} className="flex flex-col gap-4">
-                            <Input type="text" name="email" value={newTicketData.reporter_email} required disabled placeholder="Reporter email" />
-                            <Textarea
-                                name="description"
-                                id="description"
-                                required
-                                minLength={10}
-                                maxLength={250}
-                                placeholder="Ticket description"
-                                onChange={(e) =>
-                                    setNewTicketData((prev) => ({
-                                        ...prev,
-                                        description: e.target.value,
-                                    }))
-                                }
-                                value={newTicketData.description}
-                            />
-                            <Input
-                                type="file"
-                                multiple
-                                accept="image/png, image/jpeg, image/jpg"
-                                onChange={(e) => {
-                                    // const pictures = { pictures: };
-                                    setNewTicketData((prev) => ({
-                                        ...prev,
-                                        pictures: e.target.files,
-                                    }));
-                                }}
-                            />
-                            <div className="flex items-center gap-4">
-                                <Label htmlFor="notified">Do you want to be notified of changes ? </Label>
-                                <Checkbox
-                                    id="notified"
-                                    checked={newTicketData.being_notified}
-                                    onClick={() => {
-                                        setNewTicketData((prev) => ({
-                                            ...prev,
-                                            being_notified: !newTicketData.being_notified,
-                                        }));
-                                    }}
-                                />
-                            </div>
-                            <Button>Add new ticket</Button>
-                            <Button onClick={closeModalTicket} type="button" variant={'secondary'}>
-                                Cancel
-                            </Button>
-                        </form>
-                    </div>
-                </div>
-            </div>
-        );
-    };
-
     const restoreAsset = (asset: Asset) => {
         post(route('api.tenant.assets.restore', asset.id));
     };
@@ -185,7 +28,6 @@ export default function ShowAsset({ asset }: { asset: Asset }) {
         destroy(route(`api.tenant.assets.force`, asset.id));
     };
 
-    console.log(asset);
     return (
         <AppLayout breadcrumbs={breadcrumbs}>
             <Head title={`Asset ${asset.maintainable.name}`} />
@@ -209,14 +51,6 @@ export default function ShowAsset({ asset }: { asset: Asset }) {
                             <Button onClick={() => deleteAsset(asset)} variant={'destructive'}>
                                 Delete
                             </Button>
-                            <Button
-                                onClick={() => {
-                                    setSubmitTypeTicket('new');
-                                    setAddTicketModal(!addTicketModal);
-                                }}
-                            >
-                                Add new ticket
-                            </Button>
                         </>
                     )}
 
@@ -237,56 +71,7 @@ export default function ShowAsset({ asset }: { asset: Asset }) {
                 <p>Model : {asset.model}</p>
                 <p>Serial number : {asset.serial_number}</p>
 
-                <details>
-                    <summary className="">
-                        <h3 className="inline">Tickets ({tickets?.length ?? 0})</h3>
-                    </summary>
-                    {tickets.length > 0 && (
-                        <Table>
-                            <TableHead>
-                                <TableHeadRow>
-                                    <TableHeadData>Code</TableHeadData>
-                                    <TableHeadData>Status</TableHeadData>
-                                    <TableHeadData>Reporter</TableHeadData>
-                                    <TableHeadData>Description</TableHeadData>
-                                    <TableHeadData>Created at</TableHeadData>
-                                    <TableHeadData>Updated at</TableHeadData>
-                                    <TableHeadData></TableHeadData>
-                                </TableHeadRow>
-                            </TableHead>
-                            <TableBody>
-                                {tickets.map((ticket, index) => {
-                                    return (
-                                        <TableBodyRow key={index}>
-                                            <TableBodyData>
-                                                <a href={route('tenant.tickets.show', ticket.id)}>{ticket.code}</a>
-                                            </TableBodyData>
-                                            <TableBodyData>{ticket.status}</TableBodyData>
-                                            <TableBodyData>{ticket.code}</TableBodyData>
-                                            <TableBodyData>{ticket.description}</TableBodyData>
-                                            <TableBodyData>{ticket.created_at}</TableBodyData>
-                                            <TableBodyData>{ticket.updated_at !== ticket.created_at ? ticket.updated_at : '-'}</TableBodyData>
-                                            <TableBodyData>PICTURES</TableBodyData>
-
-                                            <TableBodyData>
-                                                {ticket.status !== 'closed' && (
-                                                    <>
-                                                        <Button variant={'destructive'} onClick={() => closeTicket(ticket.id)}>
-                                                            Close
-                                                        </Button>
-
-                                                        <Button onClick={() => editTicket(ticket.id)}>Edit</Button>
-                                                    </>
-                                                )}
-                                            </TableBodyData>
-                                        </TableBodyRow>
-                                    );
-                                })}
-                            </TableBody>
-                        </Table>
-                    )}
-                </details>
-
+                <TicketManager itemCodeId={asset.code} getTicketsUrl={`api.assets.tickets`} locationType="assets" itemId={asset.id} />
                 <DocumentManager
                     itemCodeId={asset.code}
                     getDocumentsUrl={`api.assets.documents`}
@@ -305,7 +90,6 @@ export default function ShowAsset({ asset }: { asset: Asset }) {
             </div>
 
             {/* {addPictures && addNewPicturesModal()} */}
-            {addTicketModal && addTicket()}
         </AppLayout>
     );
 }
