+++ resources/js/pages/tenants/tickets/show.tsx
@@ -15,12 +15,9 @@ export default function ShowTicket({ ticket }: { ticket: Ticket }) {
 
     const [ticketItem, setTicketItem] = useState(ticket);
 
-    console.log(ticketItem);
-
     const fetchTicket = async () => {
         try {
             const response = await axios.get(route(`api.tickets.get`, ticket.id));
-            console.log(response.data);
             setTicketItem(response.data.data);
         } catch (error) {
             console.error('Erreur lors de la recherche :', error);
@@ -44,19 +41,26 @@ export default function ShowTicket({ ticket }: { ticket: Ticket }) {
                     Close
                 </Button>
             )}
-            <div>
-                <p>Status: {ticketItem.status}</p>
-                <p>Description : {ticketItem.description}</p>
-                <p>Reporter : {ticketItem.reporter ? ticketItem.reporter.full_name : ticketItem.reporter_email}</p>
-                <p>Closer: {ticketItem.closer?.full_name}</p>
-            </div>
-            <div className="flex gap-4">
-                {ticketItem.pictures &&
-                    ticketItem.pictures.map((picture) => {
-                        return (
-                            <img key={picture.id} src={route('api.pictures.show', picture.id)} alt="" className="aspect-square h-64 object-cover" />
-                        );
-                    })}
+            <div className="flex h-full flex-1 flex-col gap-4 rounded-xl p-4">
+                <div>
+                    <p>Status: {ticketItem.status}</p>
+                    <p>Description : {ticketItem.description}</p>
+                    <p>Reporter : {ticketItem.reporter ? ticketItem.reporter.full_name : ticketItem.reporter_email}</p>
+                    <p>Closer: {ticketItem.closer?.full_name}</p>
+                </div>
+                <div className="flex gap-4">
+                    {ticketItem.pictures &&
+                        ticketItem.pictures.map((picture) => {
+                            return (
+                                <img
+                                    key={picture.id}
+                                    src={route('api.pictures.show', picture.id)}
+                                    alt=""
+                                    className="aspect-square h-64 object-cover"
+                                />
+                            );
+                        })}
+                </div>
             </div>
         </AppLayout>
     );
